完成家长端查看学生信息，包括学生出勤情况，学生评价和学生奖惩情况；完成学生端学生信息，包括签到和师生交流；完成教师端的家校交流功能；完成学生端、教师端的教育资讯和学生活动功能。

+++ app/src/main/java/com/example/myapplication/LoginActivity.java
@@ -66,7 +66,7 @@ public class LoginActivity extends AppCompatActivity {
                 new Thread(new Runnable() {
                     @Override
                     public void run() {
-                        account_text = Integer.parseInt(account.getText().toString());
+                        account_text = Long.parseLong(account.getText().toString());
                         Log.d(Tag, account.getText().toString());
                         password_text = password.getText().toString();
                         Log.d(Tag, password_text);
